requirements.txt

fastapi
uvicorn
numpy
opencv-python
dlib-bin
python-multipart
scipy
